Revert "Merge pull request #47994 from JuliaLang/avi/improve-semi-concrete-accuracy" (#48194)
This reverts commit 03bdf15483cbe2d06526d3b46f26e5d20893f07e, reversing
changes made to a9e0545969bb76f33fe9ad9bcf52180caa1651b9.

Fixes #48089 regression

That PR causes the runtime to store an undesirable amount of garbage,
to work around a bug in the semi-concrete interpreter failing to
correctly compute the inferred result. That should be fixed in the
semi-concrete interpreter, without bloating the runtime caches with
extraneous information.