Don't bail out of inference early if effects could still be refined (#48263)
We have an early out in inference that bails if the inferred return
type of the method being called is `Any`. This makes sense in the
absence of effects, because once the rt has hit `Any`, there is
nothing new we can learn by looking at any subsequent calls.
However, in the presence of effects, we likely want to keep
going if we can prove all methods of the callsite `:foldable` as
being `:foldable` can save significant inference time down the
line if it enables concrete evaluation of the containing function.