irinterp: Remove redundant `is_inlineable_constant` check
The way that `is_inlineable_constant` is supposed to work is that during
type inference, we allow a large `Const` type to persist in the inference
domain and then at inlining time, we decide whether or not to remove the
call and replace it by the constant (thus having to serialize the constant)
or just leave the call as is (with the types eventually being widened).
`irinterp` was being a bit overeager here, not even permitting large
constants into the inference domain, which was hurting precision.
We alraedy have an appropriate `is_inlineable_constant` call guarding
the inlining of constants into the IR, so the move the one that checks
it when returning from concrete eval.