irinterp: Allow for non-Const first argument (#48326)
The first argument isn't special here in any way. The original
code here just wanted to avoid calling is_all_const_arg with
the first argument included, because that condition was already
computed when determining `f`. This code pre-dated semi-concrete
eval and is now in the wrong place.

The changes on the inlining semi-concrete interpreted opaque closure is 
also added because now we allow irinterp on opaque closure inference.

Co-authored-by: Shuhei Kadowaki <aviatesk@gmail.com>