RFC: support iteration for `CartesianIndex`
I'm almost certain this came up before, but wasn't able to find the
discussion. I think now that the multidimensional indexing API has
matured a bit, the potential for confusion when allowing iteration is
excedingly low. With this change, it's possible to destructure
`CartesianIndex` objects without having to convert them to a tuple,
which I feel currently makes writing such code unnecessarily annoying.

This also makes `CartesianIndex` broadcast like a scalar which I felt
was consistent with the behavior for objects like `Pair`, but I am open
to other suggestions.