Axes tip in `AbstractArray` interface (#47221)
Update the `AbstractArray` interface to mention that the axes of an array should, in general, be their own axes. This is required for internal consistency, and if this is not enforced (e.g. by choosing `UnitRange`s as axes instead of `Base.IdentityUnitRange`s), one may come across unexpected bugs (see https://juliaarrays.github.io/OffsetArrays.jl/stable/internals/#Wrapping-other-offset-array-types for an example).

Co-authored-by: Viral B. Shah <ViralBShah@users.noreply.github.com>