Avoid unnecessary Docs.META initializations
If the target module does not have a Docs.META dict (e.g. if
`--strip-metadata` is used), `Docs.meta()` has the side effect of
creating a new IdDict and initializing the Docs.META field of the target
module.

We need to avoid eval'ing into modules after they've been closed, so for
methods that do not mutate the new IdDict we should avoid the init.

Resolves #48390.

Co-authored-by: Steve Kelly <kd2cca@gmail.com>