More stringent range testing (#48465)
Turns a test_skip into a test. This may have been passing since
PR#23194 but never made a requirement. This change removes the
single largest contribution to "total broken tests" in the whole
test suite.

The threshold needs to be adjusted to take into account intermediate rounding in computing stop

Co-authored-by: Simon Byrne <simonbyrne@gmail.com>
Co-authored-by: Ian Butterworth <i.r.butterworth@gmail.com>