subtyping: Relax conditions for separable tuple fast-path
This check is unnecessary, since the LHS type is a subtype iff all
union-decision-paths are subtypes of the RHS type.

The RHS check needs to stay, since multiple union-decision-paths on the
RHS may have to be considered together, even when none of the individual
paths are sufficient on their own to conclude `a <: b`.