all(Array{Int}) consistent with all(Array{Int};dims=N)
Currently we have this behavior, which seem inconsistent:

```julia
julia> all([1,0];dims=1)
1-element Vector{Bool}:
 0

julia> all([1,0])
ERROR: TypeError: non-boolean (Int64) used in boolean context
```
This occurs because of the `elseif v` clauses, which fail for `v === 1`. If we instead check for `v == true`,  than it will pass, and the behavior becomes consistent among these two calls. 

I don't know if this is considered breaking (probably it is), though.

The implementation for tuples specifically had chosen the `elseif v` syntax to be consistent with `any`, such that probably this is a desired behavior. Probably this PR won't be merged, then, although the inconsistency with `any([1, 0])` will remain.