Add Julia side code
This is almost entirely copying SimpleVector implementations. When
calling `Core.sbuf(...)` it still throws a segmentation fault 11. Right
now my only guess is that I inadvertently have miscounted the amount of
builtin tags that I inserted and that it's trying to access memory that
isn't there. I'm still figuring out that entire bit of code though, so
it's easier said than done.