allow external `AbstractInterpreter` to keep source of const-folded cache (#48626)
After #48598, the `CodeInstance` constructor discards `inferred_result`
when the cache can use the constant calling convention.
This is generally fine, but external `AbstractInterpreter` may still want
to keep the discarded source, that may be arbitrary custom data
structure. This commit makes use of the `may_discard_trees` interface to
configure that behavior.