Avoid creating :invoke for unbound tparams
A `MethodInstance` whose static parameter values are unknown will
have `TypeVar`s in the corresponding slot in their sparam_vals
object and have their code instance's `->invoke` method set to
`jl_fptr_sparam`. I think the original expectation here was
that at some point the runtime would externally compute the
set of static parameters and pass them in as an argument
(just as it would for regular arguments). However, as far as
I can tell, no place in the runtime actually does this and
instead static paramters are always tied to a particular
MethodInstance. This is enforced by making sure that
compilable signatures never have unbound typevars (unless
that is the true answer). However, when we added the
`compilesig_invokes` optimizer option, this allowed bypassing
`get_compileable_sig`, causing unexpected errors around
type parameters.

This commit simply institutes a check that fixes this
particular case, though I don't think the idea of wanting
to :invoke MethodInstances with unknown sparams is
all that unreasonable (especially since we can now
inline such cases) and in the future we may want to
revisit the runtime support for actually passing through
sparams.