effects: fix unknown effects
The `noinbounds` bit of `EFFECTS_UNKNOWN` should be `false` since
arbitrary method may have `:boundscheck`.

This commit also replaces all the usages of `EFFECTS_UNKNOWN′` by
`Effects()` and renames it to `_EFFECTS_UNKNOWN` to indicate that
it is very internal and is supposed to be used by the `Effects(...)`
constructor only.