Setup `pthread_sigmask` before loading libjulia-internal
Loading libjulia-internal and libjulia-codegen can trigger constructors
in their dependencies, including some which may spawn threads (e.g.
Tracy). It is important that these not be configured to receive any
signals that libjulia-internal is expecting to `sigwait()` on, esp. when
their default action is to terminate the process.

Technically, the masked signal set should probably be synchronized with
the `sigwait_sigs` set in `signals-unix.c`, but for now this simply masks
all signals.