Add an inference option to ignore the recursion hardlimit
We currently very aggressively limit recursion if we came from
a union split. The reason for this choice is to avoid accidental
exponential inference times from excessive exploration of the call
graph. Unfortunately, certain packages like Diffractor really like
triggering the recursion heuristic at the moment, causing any unions
to immediately cause imprecise inference. In the fullness of time,
we should improve the recursion heuristic to do better in this case,
but for the time being, add an inference option that simply lets
it ignore the hardlimit recursion case and proceed anyway.