irinterp: Allow setting all IR flags
Currently, IR_FLAG_NOTHROW is the only flag that irinterp is allowed to
set on statements, under the assumption that in order for a call to
be irinterp-eligible, it must have been proven :foldable, thus :effect_free,
and thus IR_FLAG_EFFECT_FREE was assumed to have been set. That reasoning
was sound at the time this code was written, but have since introduced
EFFECT_FREE_IF_INACCESSIBLEMEMONLY, which breaks the reasoning that
an :effect_free inference for the whole function implies the flag on
every statement. As a result, we were failing to DCE otherwise dead
statements if the IR came from irinterp.