Lookup `ccall` symbols in local internal libraries first
This is similar to what we do on Windows, where handles to the exact
library are required anyway. This is enough to make sure that our
`dlsym` lookups are directed to the correct libjulia, even when loading
a Julia run-time within Julia.

The second change needed to get things working (not included in this
commit) is to add symbol versioning, so that the runtime linker does not
mix up symbols between the two libraries.