Makefile: Add support for symbol versioning (off by default)
This change adds `DEFAULT_SYMBOL_VERSION` as a Makefile option. If
specified, this will configure the linker to assign the specified
version string to all Julia-exported symbols.

The other half of the mechanism is `#define JL_SYMBOL_VERSION`, which
is used by the ccall machinery to look up symbols by their version + name.