datatype: handle concrete type intersections that happen
This is actually very similar to the current might_intersect_concrete,
but for subtyping and memoized. It replaces cached_by_hash, which was a
confusingly-named incomplete prior work towards this.

This gives NamedTuple{(:names)} hashes, which lets them go into the
faster type lookup tables. This is a fairly common type for some
packages to create, so we need this to avoid polluting our cache tables.

Reverts efafd8388675d65096e0f088ddfe96f4e8077567, since these types have
no intersection, the morespecific algorithm is no longer required to
have any opinion on them.