Set various array and Random effects
My primary motivation here was to get the Xoshiro constructor to
be removed if the RNG is unused (I have a struct that always has
an RNG in it, which may or may not be used). I first tried to
do that in a more fine-grained way, but #47595 currently makes
that infeasible. Nevertheless, I figured it was worth leaving
those improvements in so we can switch over once #47595 is
resolved. In the meantime, annotate the top-level Xoshiro
constructor as `:removable`.