Fix emitting `undef` when coercing an `Int` to the pointersize of a target
Previously, we always emitted `undef` in our IR when the bitsize of
an unboxed value and the pointersize we want to coerce it to were
mismatched. This is an issue when the pointersize is smaller than
a supported integer size of a platform. This commit fixes this,
by only emitting the `undef` when both arguments `to` and `ty` are
not integers and not pointers, which we should be fine with converting
between.