profiling: Add Task/GC/Inference/Codegen signposts for Tracy
This adds a variety of metadata to send to Tracy during profiling, in
particular method instance signatures, the Heap size, and a couple of
other small pieces of relevant metadata.

It also adds full support for Task execution, tracking these as "Fibers"
on the Tracy side. This does incur some overhead until an enhancement
is implemented on the Tracy-side to avoid a globally synchronous Fiber
event queue, but most of the events that we're tracking now are
relatively coarse so it should not be a major issue.