Add retry for socket race condition in libgit2 tests (#49089)
We have a race condition in the libgit2 tests where we're probing for
an open port and then try to acquire it later. However, that leaves
a window for a different process to grab the open port in the meantime
and we regularly see this failure on CI. A proper fix would be to let
s_server do its own probing, add the ability to pass in a file descriptor
rather than a port number or switch to a different SSL server, but
absent somebody wanting to that work, this does the next best thing
of just retrying the server start.