rewrite DW_TAG_compile_unit during aotcompile
The macOS linker requires these to be unique due to a bug, and we do not
care about this value at all as it does nothing, so just make it
something unique. However, keep the directory as ".", since it uses that
to compose the DW_AT_decl_file values for its subprograms.

Also try to save a little memory, since we have to leak all of these
objects (per LLVMContext), and we would like to avoid that.

Fix #49152
Fix #49151
Fix #49153