Simplify child layout for `try` (JuliaLang/JuliaSyntax.jl#234)
The child layout of try-catch-else-finally is awkward because several of
the subclauses are optional. Particularly this has led to problems for
`else` in `Expr` which needed to be tacked onto the end for
compatibility.

This change clarifies the situation a bit and makes it more future proof
by wrapping the subclauses in their own expression heads.