profiling: Add `WITH_TIMING_COUNTS` build option
This replaces the `ENABLE_TIMINGS` #define in `options.h`

Since we now support multiple back-ends for timing events, each back-end
is expected to be enabled with a `WITH_*` Makefile flag, and then the
JL_TIMING API is automatically enabled if any back-end is available.

Turning on multiple back-ends at the same time is also supported.