normalize unions made in subtyping
We observed a case where simple_tmeet made a Union of egal things, which
is undesirable. There also was no sorting of the result, as it normally
done, and theoretically, simplification with an omit_bad_union to remove
`S` could similar result in a Union that should be further simplified to
remove redundancies.
```
Union{Union{Val{T}, S} where T<:AbstractString, Union{Val{T}, Int64} where T<:AbstractString} where S
```
(In principle, that simplification might also be possible to do during
the original jl_type_union call when flattening it.)