subtype: reuse existing allocations when possible
When part of a type is a subtype of the other part, we can reuse the
existing object to represent the result of that part. Additionally, we
can sometimes defer allocating this until it is known how long it needs
to be and whether it is already present in the type cache.