API updates: `parse()` -> `parsestmt()` + exports + `parse!(rule=:all)`
`JuliaSyntax.parse()` clashes with `Base.parse()` and it's really not clear
whether `parse()` should parse a single statement or a whole file
top-level. Having a less generic name `parsestmt()` helps with this.

Using the name `parsestmt()` is explicit about the fact that this parses a
single statement (not an "expression" which is ambiguous given
every Julia construct is an expression). It's also consistent with the
naming rules of `parseall()` and `parseatom()`, which already appear in
Base as `Meta.parseall` and `Meta.parseatom`.

Change to using `rule=:all` in the `parse!()` API rather than `rule=:toplevel`
because this is most consistent with the choices made in the `Core._parser`
interface and the naming of `parseall()`.

Also add a conservative list of exports that I expect "people are likely
to use", and which seem to be required as the main part of the API.
There's more to the API than this, but shoving that all into
the user's namespace doesn't seem ideal. Especially the parts which
are less certain.