Prettify printing of a small number of completions
Previously show_completions() tended to maximize the number of columns used,
and if there were few enough completions to only use one or two lines the text
would be a little difficult to read. Now it tries to minimize the number of
columns, and there's a minimum number of completions needed before it'll start
using multiple columns.