sroa_pass!: Mark statements that were updated
sroa_pass! can introduce new type information that inference did not
see (e.g. when forwarding field accesses through a mutable). Currently,
we just accept that in base, but there are external pass pipelines
that run multiple rounds of inference and would like to refine these.
Make this possible by introducing a new IR flag that sroa_pass! sets
when it modifies a statement. There are currently no consumers in the
Base pipeline, but the idea is that these can be passed into
_ir_abstract_constant_propagation for refinement.