Fix TODO in IncrementalCompact's `already_inserted`
The `already_inserted` query on `IncrementalCompact` had a missing
case for being called on incremental compact whose underlying IR
had new nodes to be inserted. This does not happen in the base pipeline,
because `already_inserted` is only used in the sroa passes, but can
happen in some non-Base pipelines that run multiple rounds of sroa.