Always parse standalone dotted operators as `(. op)` (JuliaLang/JuliaSyntax.jl#240)
Change the parsing of standalone dotted operators such as `.+` to always
parse as `(. +)`, regardless of where they appear syntactically. This
removes the last of the cases where operators were coalesced with their
leading `.` into a single symbol. For example, `.+(x,y)` is now parsed
as `(call (. +) x y)`, rather than `(call .+ x y)`.

The reference parser concatenates the `.` and the operator in some
cases, so for compatibility we convert back to that less consistent form
during `Expr` conversion.